# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Race Analysis Engine",
    "description": "\nAdvanced Formula 1 analytics.\n\nEvery analysis response carries the `engine_version` that produced it. Cached results\nare keyed on that version, so changing a model invalidates its cached values rather\nthan mixing two definitions in one chart.\n\n**On estimates.** Degradation slopes, fuel-corrected lap times and pit loss are modelled\nfrom observed timing, not measured from team telemetry. Fields that carry an estimate\nsay so in their description. In particular the fuel coefficient is a published\n0.030 s/kg prior, not a value fitted from this dataset \u2014 it cannot be fitted from race\ndata, because fuel load is derived from lap number and carries no independent information.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis/laps/{session_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Laps",
        "description": "Per-lap metrics, including why any lap was excluded.",
        "operationId": "get_laps_api_v1_analysis_laps__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "driver",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one car number",
              "title": "Driver"
            },
            "description": "Restrict to one car number"
          },
          {
            "name": "representative_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only laps that belong in a pace sample",
              "default": false,
              "title": "Representative Only"
            },
            "description": "Only laps that belong in a pace sample"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LapOut"
                  },
                  "title": "Response Get Laps Api V1 Analysis Laps  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/pace/{session_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Pace",
        "description": "Driver pace, ranked.\n\nPace is the 20th percentile of clean fuel-corrected laps \u2014 not the fastest lap,\nwhich rewards whoever had the best single opportunity, and not the mean, which\nrewards whoever avoided traffic.",
        "operationId": "get_pace_api_v1_analysis_pace__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/degradation/{session_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Degradation",
        "description": "Tyre degradation by compound, with the stint fits behind it.\n\nA stint fit may report a negative slope when the run was too short to support an\nestimate. Those are returned with ``is_physical`` false rather than hidden \u2014 the\nfailure is information about the stint.",
        "operationId": "get_degradation_api_v1_analysis_degradation__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DegradationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DegradationOut": {
        "properties": {
          "compound": {
            "type": "string",
            "title": "Compound"
          },
          "n_stints": {
            "type": "integer",
            "title": "N Stints"
          },
          "n_laps": {
            "type": "integer",
            "title": "N Laps"
          },
          "degradation_s_per_lap": {
            "type": "number",
            "title": "Degradation S Per Lap",
            "description": "Median slope across reliable stints on this compound"
          },
          "degradation_iqr_s": {
            "type": "number",
            "title": "Degradation Iqr S",
            "description": "Interquartile spread: the honest width of the estimate"
          },
          "median_pace_s": {
            "type": "number",
            "title": "Median Pace S"
          },
          "max_stint_laps": {
            "type": "integer",
            "title": "Max Stint Laps",
            "description": "Longest observed stint. Predictions beyond this extrapolate."
          }
        },
        "type": "object",
        "required": [
          "compound",
          "n_stints",
          "n_laps",
          "degradation_s_per_lap",
          "degradation_iqr_s",
          "median_pace_s",
          "max_stint_laps"
        ],
        "title": "DegradationOut"
      },
      "DegradationResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "compounds": {
            "items": {
              "$ref": "#/components/schemas/DegradationOut"
            },
            "type": "array",
            "title": "Compounds"
          },
          "stints": {
            "items": {
              "$ref": "#/components/schemas/StintFitOut"
            },
            "type": "array",
            "title": "Stints"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "meta",
          "compounds",
          "stints"
        ],
        "title": "DegradationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LapOut": {
        "properties": {
          "driver_number": {
            "type": "string",
            "title": "Driver Number"
          },
          "lap_number": {
            "type": "integer",
            "title": "Lap Number"
          },
          "lap_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lap Time S"
          },
          "compound": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound"
          },
          "tyre_life": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tyre Life"
          },
          "stint": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stint"
          },
          "position": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "is_representative": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Representative",
            "description": "Whether this lap belongs in a pace sample"
          },
          "exclusion_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusion Reason",
            "description": "Why the lap was excluded, when it was"
          },
          "fuel_corrected_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Corrected S",
            "description": "Lap time corrected to an empty-tank equivalent using a published 0.030 s/kg coefficient. An estimate, not measured fuel telemetry."
          }
        },
        "type": "object",
        "required": [
          "driver_number",
          "lap_number"
        ],
        "title": "LapOut"
      },
      "Meta": {
        "properties": {
          "engine_version": {
            "type": "string",
            "title": "Engine Version",
            "description": "Engine version that produced these values"
          },
          "computed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed At",
            "description": "When the underlying metrics were computed"
          }
        },
        "type": "object",
        "required": [
          "engine_version"
        ],
        "title": "Meta",
        "description": "Provenance attached to every analysis response."
      },
      "PaceOut": {
        "properties": {
          "driver_number": {
            "type": "string",
            "title": "Driver Number"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "n_laps": {
            "type": "integer",
            "title": "N Laps"
          },
          "finish_position": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Position",
            "description": "Classified finishing position, null if not classified"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Finished, Lapped, Retired, or a stated cause"
          },
          "laps_completed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Laps Completed"
          },
          "did_not_finish": {
            "type": "boolean",
            "title": "Did Not Finish",
            "description": "True when the driver retired. Their pace still counts \u2014 it is measured from the laps they did complete \u2014 but the result did not follow it.",
            "default": false
          },
          "pace_s": {
            "type": "number",
            "title": "Pace S",
            "description": "20th percentile of clean fuel-corrected laps, not the fastest lap"
          },
          "gap_to_best_s": {
            "type": "number",
            "title": "Gap To Best S"
          },
          "best_s": {
            "type": "number",
            "title": "Best S"
          },
          "median_s": {
            "type": "number",
            "title": "Median S"
          },
          "std_s": {
            "type": "number",
            "title": "Std S",
            "description": "Lap-time spread: lower is more consistent"
          },
          "clean_air_laps": {
            "type": "integer",
            "title": "Clean Air Laps"
          }
        },
        "type": "object",
        "required": [
          "driver_number",
          "rank",
          "n_laps",
          "pace_s",
          "gap_to_best_s",
          "best_s",
          "median_s",
          "std_s",
          "clean_air_laps"
        ],
        "title": "PaceOut"
      },
      "PaceResponse": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "drivers": {
            "items": {
              "$ref": "#/components/schemas/PaceOut"
            },
            "type": "array",
            "title": "Drivers"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "meta",
          "drivers"
        ],
        "title": "PaceResponse"
      },
...
      "StintFitOut": {
        "properties": {
          "driver_number": {
            "type": "string",
            "title": "Driver Number"
          },
          "stint": {
            "type": "integer",
            "title": "Stint"
          },
          "compound": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound"
          },
          "n_laps": {
            "type": "integer",
            "title": "N Laps"
          },
          "pace_s": {
            "type": "number",
            "title": "Pace S",
            "description": "Fitted lap time at zero tyre age"
          },
          "degradation_s_per_lap": {
            "type": "number",
            "title": "Degradation S Per Lap",
            "description": "Fitted slope. May be negative when the stint was too short to support an estimate; see is_physical."
          },
          "is_physical": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Physical",
            "description": "False when the fitted slope is negative. Tyres do not gain time with age, so such a fit means the stint gave too little usable range."
          },
          "is_reliable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Reliable"
          },
          "r_squared": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R Squared"
          },
          "tyre_age_range": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tyre Age Range",
            "description": "Span of tyre age the fit actually covered"
          },
          "excluded_lap_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Lap Count",
            "description": "Laps the warm-up cutoff removed before fitting"
          }
        },
        "type": "object",
        "required": [
          "driver_number",
          "stint",
          "n_laps",
          "pace_s",
          "degradation_s_per_lap"
        ],
        "title": "StintFitOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}